# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Accounting API Documentation",
        "description": "This is the API documentation for the Accounting Application.",
        "contact": {
            "name": "API Support",
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/reports/transactions": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get transactions report data",
                "description": "Returns transaction data for reports based on filters",
                "operationId": "c3ebdd7fdfeb248f21fb53f90c0e2646",
                "parameters": [
                    {
                        "name": "date_range",
                        "in": "query",
                        "description": "Predefined date range",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "this_month",
                                "last_month",
                                "last_3_months",
                                "last_6_months",
                                "this_year",
                                "custom"
                            ]
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Start date for custom range (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "End date for custom range (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "budget_id",
                        "in": "query",
                        "description": "Filter by specific budget ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "description": {
                                                        "type": "string"
                                                    },
                                                    "amount": {
                                                        "type": "number"
                                                    },
                                                    "transaction_date": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "budget": {
                                                        "type": "object"
                                                    },
                                                    "sub_user": {
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/reports/summary": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get summary report data",
                "description": "Returns summary statistics for reports",
                "operationId": "1f2e431ccb10dff5dc59f426d479811c",
                "parameters": [
                    {
                        "name": "date_range",
                        "in": "query",
                        "description": "Predefined date range",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "this_month",
                                "last_month",
                                "last_3_months",
                                "last_6_months",
                                "this_year",
                                "custom"
                            ]
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Start date for custom range (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "End date for custom range (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "total_expenses": {
                                                    "type": "number"
                                                },
                                                "total_income": {
                                                    "type": "number"
                                                },
                                                "net_balance": {
                                                    "type": "number"
                                                },
                                                "transaction_count": {
                                                    "type": "integer"
                                                },
                                                "by_budget": {
                                                    "type": "object"
                                                },
                                                "by_month": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use a JWT token to access these endpoints. Format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}